Integrate S3 Storage Backend (#919)
* design library

* integrate backpack in file cache

* replace direct fs cals with cache

* fix tests

* clean code

* start s3

* dummy implementation

* fix file browsing

* add remove dir

* improve synchronization

* fix block syncronization

* remove uncessary rwlock

* clean code

* handle access mode in file

* integrate local cache quota

* fix sync files after download

* fix dead lock in keeping quota

* add debug message for removing

* compile tests

* fix tests

* clean code

* bump  version

* bring backend to reductstore

* remove unused methods

* fix version test

* coverage without plugins and console

* refactor s3 connector

* isolate backen features

* add unit tests for local cache

* test s3 connector

* add tests for remote backend

* add tests for file system backend

* fix windows bug

* add tests for backend

* add test for file

* test backend builder

* add tests for config

* fix config test

* fix for aws

* fix test

* test s3 connector in ci

* enable ci in coverage

* use minio in ci

* fix variable names

* add dirs

* fix syntax

* deploy minio manually

* fix bucket env

* use actions to run minio

* fix container name

* forgot checkout

* test ci only with coverage

* test recursive listing

* fix recovery test

* add rename test

* update CHANGELOG

* copilot review

+++ reductstore/src/api/entry/write_batched.rs
@@ -199,15 +199,15 @@ async fn write_chunk(
     writer: &mut Box<dyn WriteRecord + Sync + Send>,
     chunk: Bytes,
     written: &mut usize,
-    content_size: usize,
+    content_size: u64,
 ) -> Result<Option<Bytes>, ReductError> {
-    let to_write = content_size - *written;
+    let to_write = content_size - *written as u64;
     *written += chunk.len();
-    let (chunk, rest) = if chunk.len() < to_write {
+    let (chunk, rest) = if (chunk.len() as u64) < to_write {
         (chunk, None)
     } else {
-        let chuck_to_write = chunk.slice(0..to_write);
-        (chuck_to_write, Some(chunk.slice(to_write..)))
+        let chuck_to_write = chunk.slice(0..to_write as usize);
+        (chuck_to_write, Some(chunk.slice(to_write as usize..)))
     };
 
     writer
@@ -220,7 +220,7 @@ fn check_content_length(
     headers: &HeaderMap,
     timed_headers: &Vec<(u64, RecordHeader)>,
 ) -> Result<(), ReductError> {
-    let total_content_length: usize = timed_headers
+    let total_content_length: u64 = timed_headers
         .iter()
         .map(|(_, header)| header.content_length)
         .sum();
@@ -231,7 +231,7 @@ fn check_content_length(
             .ok_or(unprocessable_entity!("content-length header is required",))?
             .to_str()
             .unwrap()
-            .parse::<usize>()
+            .parse::<u64>()
             .map_err(|_| unprocessable_entity!("Invalid content-length header"))?
     {
         return Err(unprocessable_entity!(
